Add recommended speed feature to running exercise configuration
- Introduced `recommendedSpeed` property in ExerciseSegmentDTO and ExerciseSegmentEntity.
- Updated mock data to include recommended pace.
- Added RecommendedSpeedHint widget to display recommended speed during active sessions.
- Adjusted relevant tests to accommodate new property.

+++ lib/features/running/domain/entities/running_exercise_config.dart
@@ -28,7 +28,7 @@ class RunningExerciseConfig {
               metric: segment.targetMetric,
               limitValue: segment.targetMetric == WorkoutMetric.time
                   ? segment.durationSec.toDouble()
-                  : segment.distanceM.toDouble(),
+                  : segment.distanceM,
               segmentId: segment.id,
               activityType: segment.activity,
             ),
